+++ CHANGELOG.md
@@ -23,6 +23,49 @@ the code was deployed.
 - Added migration script to change the default value in the `incident_categories` column to `{"Overdose Event","Safer Use","Conflict Resolution","Test"}` (CU-86du7hv9b)
 - Added migration script to insert `organization`, `funder`, and `postal_code` columns into the `clients_extention` table (CU-86du6jp33).
 - New REST API routes (still require test cases before further use) (CU-86dugzyzh).
+- Added `ClientExtension.js` with class `ClientExtension` to be used in adding and editing clients (CU-86du23zth).
+- Added device dashboard page with `locationsDashboard.mst` mustache file (CU-86du23zth).
+- Added new CSS for the device dashboard page in `locationsCSSPartial.mst` (CU-86du23zth).
+- Added update client page with `updateClient.mst` mustache file (CU-86du23zth).
+- Added `renderUpdateClient` function in `dashboard.js` to render the mustache file (CU-86du23zth).
+- Added `validateUpdateClient` function in `dashboard.js` to validate the update client form (CU-86du23zth).
+- Added `submitUpdateClient` function in `dashboard.js` to submit the update client form (CU-86du23zth).
+- Added new client page with `newClient.mst` mustache file.
+- Added `renderNewClientPage` function in `dashboard.js` to render the mustache file (CU-86du23zth).
+- Added `validateNewClient` function in `dashboard.js` to validate the new client form (CU-86du23zth).
+- Added `submitNewClient` function in `dashboard.js` to submit the new client form (CU-86du23zth).
+- Added update gateway page with `updateGateway.mst` mustache file (CU-86du23zth).
+- Added `renderUpdateGatewayPage` function in `dashboard.js` to render the mustache file (CU-86du23zth).
+- Added `validateUpdateGateway` function in `dashboard.js` to validate the update gateway form (CU-86du23zth).
+- Added `submitUpdateGateway` function in `dashboard.js` to submit the update gateway form (CU-86du23zth).
+- Added new gateway page with `newGateway.mst` mustache file (CU-86du23zth).
+- Added `renderNewGatewayPage` function in `dashboard.js` to render the mustache file (CU-86du23zth).
+- Added `validateNewGateway` function in `dashboard.js` to validate the new gateway form (CU-86du23zth).
+- Added `submitNewGateway` function in `dashboard.js` to submit the new gateway form (CU-86du23zth).
+- Added update button page with `updateButton.mst` mustache file (CU-86du23zth).
+- Added `renderUpdateButtonPage` function in `dashboard.js` to render the mustache file (CU-86du23zth).
+- Added `validateUpdateButton` function in `dashboard.js` to validate the update button form (CU-86du23zth).
+- Added `submitUpdateButton` function in `dashboard.js` to submit the update button form (CU-86du23zth).
+- Added `buttonFormCSSPartial.mst` that has all the CSS for every form page in the dashboard (CU-86du23zth).
+- Added `New Clients` and `New Gateway` buttons to the nav bar (CU-86du23zth).
+- Added `Edit Gateway` buttons to the vitals page next to each gateway (CU-86du23zth).
+- Added several new routing to route pages into each other in `routes.js` (CU-86du23zth).
+- Added function `createClientExtensionFromRow` in `db,js` for new pages (CU-86du23zth).
+- Added function `getGatewayWithGatewayId` in `db,js` for new pages (CU-86du23zth).
+- Added function `getHistoryOfSessions` in `db,js` for new pages (CU-86du23zth).
+- Added function `getButtonWithDeviceId` in `db,js` for new pages (CU-86du23zth).
+- Added function `getClientExtensionWithClientId` in `db,js` for new pages (CU-86du23zth).
+- Added function `getMostRecentSessionWithDevice` in `db,js` for new pages (CU-86du23zth).
+- Added function `updateClient` in `db,js` for new pages (CU-86du23zth).
+- Added function `createClientExtension` in `db,js` for new pages (CU-86du23zth).
+- Added function `updateClientExtension` in `db,js` for new pages (CU-86du23zth).
+- Added function `createGatewayFromBrowserForm` in `db,js` for new pages (CU-86du23zth).
+- Added function `updateButton` in `db,js` for new pages (CU-86du23zth).
+- Added function `updateGateway` in `db,js` for new pages (CU-86du23zth).
+
+### Changed
+- Changed `clientPage.mst` to show all current client buttons instead of all vitals (CU-86du23zth).
+- Changed `renderButtonDetailsPage` function in `dashboard.js` to reflect new page changes for the client page (CU-86du23zth).
 
 ## [13.11.0] - 2024-08-20